feat(FVC-151): implement websocket

+++ fvc-api/src/main/java/sep490g65/fvcapi/entity/EvaluationCriteria.java
@@ -1,6 +1,7 @@
 package sep490g65.fvcapi.entity;
 
 import jakarta.persistence.*;
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -21,7 +22,7 @@ public class EvaluationCriteria extends BaseEntity {
     private String name;
 
     @Column(precision = 5, scale = 2, nullable = false)
-    private Double weight; // should sum to 100 per session template; enforced by service
+    private BigDecimal weight; // should sum to 100 per session template; enforced by service
 }
