+++ fvc-api/src/main/java/sep490g65/fvcapi/entity/EvaluationScore.java
@@ -1,6 +1,7 @@
 package sep490g65.fvcapi.entity;
 
 import jakarta.persistence.*;
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -28,7 +29,7 @@ public class EvaluationScore extends BaseEntity {
     private EvaluationCriteria criteria;
 
     @Column(precision = 5, scale = 2, nullable = false)
-    private Double score;
+    private BigDecimal score;
 }
